ml/requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
tensorflow==2.17.1
pillow==10.4.0
pyyaml==6.0.2
joblib==1.4.2
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.12
